Apisecurity

+++ controllers/playersController.js
@@ -50,6 +50,10 @@ module.exports = app => {
 
   // creates player
   app.post('/api/players', (req, res) => {
+    if(req.body.authKey !== process.env.AUTH_KEY) {
+      res.status(403).send( {'error' : "bad auth key"});
+      return;
+    }
     var newPlayer = Players(req.body);
     newPlayer.save((error, player) => {
       error ? res.status(501).send({ error }) : res.send(player);
@@ -59,6 +63,10 @@ module.exports = app => {
 
   // updates player
   app.put('/api/players/:id', (req, res) => {
+    if(req.body.authKey !== process.env.AUTH_KEY) {
+      res.status(403).send( {'error' : "bad auth key"});
+      return;
+    }
     Players.findByIdAndUpdate(req.params.id, req.body, (error, player) => {
       error ? res.status(501).send({ error }) : res.send(player);
       players_cache.force_reload();
@@ -67,6 +75,10 @@ module.exports = app => {
 
   //deletes player
   app.delete('/api/players/:id', (req, res) => {
+    if(req.body.authKey !== process.env.AUTH_KEY) {
+      res.status(403).send( {'error' : "bad auth key"});
+      return;
+    }
     Players.findByIdAndRemove(req.params.id, error => {
       error
         ? res.status(501).send({ error })
